Dragging: Move drag event handlers to the top level template.
Previously the drag event handlers were attached directly to the
hexes template. This worked well for dragging that stayed within
the hexes, but if the mouse wandered or the hexes hadn't loaded
correctly it caused the dragging to be interrupted.

Additionally, the dragging was cancelled whenever the mouse left
the dragged element. There is no reason dragging shouldn't be
allowed to continue when the mouse leaves the element, care simply
needs to be taken with event propagation.