notifications_global: filter out multiple new_dominus notifications
Uses the fact that each user has their own copy of the notification.
This should almost certainly be reworked to do one of
a. Remove the notification from the global panel
b. Create a new 'new_dominus_global' notification in place of reusing
   the existing personal notifications
c. Do the filtering on the server side